requirements.txt
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
python-telegram-bot[job-queue]>=20.7
anthropic>=0.18.0
google-genai>=1.0.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
flask>=2.3.0
gunicorn>=21.2.0
pytz>=2024.1
lxml>=5.0.0
